+++ src/util/ResourceLocator.hpp
@@ -2,5 +2,6 @@
 #include <filesystem>
 namespace pkmn::cli::util {
 std::filesystem::path RedTemplatePath();
+std::filesystem::path FireRedTemplatePath();
 std::filesystem::path FireRedRuntimeScriptPath();
 } // namespace pkmn::cli::util